fix: correctly manage popover creation/destroy
The popover weren't displayed anymore after we allow to display the content of the subprocess. This
is now fixed. In addition, configure 'tippy's sticky' settings correctly.

Update the cursor on the activity with 'compliance rules' violations to state that the activity is
clickable.
Manage the tippy instances (destroy on hide) and visual elements lifecycle.

In addition
  - center main diagram at load time
  - remove dead code
  - remove console log

Notice there are a lot of refactoring needed in the file. It will be managed later.