[FIX] The reset button also resets the state of the other buttons (#46)
With the last version, if we showed one or more data, the data were reset on click on the `Reset` button, but not the state of the buttons and their label. 
Because of that, when we tried to click again on those buttons the first time, the data were not displayed.