feat(happy-path): highlight elements less linearly (#69)
Alternate the branch while highlighting to better reflect the execution of the process because elements of various branches can be executed in parallel. Also remove some ids that were added twice in the list of the
elements used to highlight the path.

Bonus: introduce speed factor to accelerate the animation while debugging.